Update for new Jasper plugin system
Refactored to new plugin class
Switched to using 'requests' instead of 'urllib2'
